Restructure project to src layout and update config
- Move all app and test code under src/ (src/app, src/tests)
- Remove backend/ directory and related files
- Update imports for new src/app structure and relative imports
- Rename project to "tiny-sales" in pyproject.toml
- Update CLI and test paths for src layout
- Add per-feature conftest.py files for test fixtures
- Update dependencies and lock files for new project name and latest versions
