+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "backend"
+name = "tiny-sales"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
@@ -13,15 +13,25 @@ dependencies = [
     "aerich[toml]>=0.8.0",
     "python-jose[cryptography]>=3.5.0",
     "bcrypt>=4.3.0",
-    "typer[all]>=0.12.3", # Added Typer
 ]
 
+[tool.uv]
+package = true
+
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
 [project.scripts]
-manage-users = "backend.cli.main:app"
+manage-users = "app.cli.main:app"
 
 [tool.aerich]
 tortoise_orm = "app.main.TORTOISE_ORM_CONFIG"
 location = "./migrations"
 
 [dependency-groups]
-dev = ["pytest>=8.4.0", "pytest-asyncio>=1.0.0", "pytest-httpx>=0.21.2"] # httpx is already installed, this adds the pytest plugin
+dev = [
+    "pytest>=8.4.0",
+    "pytest-asyncio>=1.0.0",
+    "pytest-httpx>=0.21.2",
+] # httpx is already installed, this adds the pytest plugin
