+++ pytest.ini
@@ -1,0 +1,8 @@
+# /tiny-sales/pytest.ini
+[pytest]
+pythonpath = src
+testpaths =
+    src/tests
+    src/app/features/reports/tests
+    src/app/features/orders/tests  # If your tests are here, add this line
+norecursedirs = .venv .* build dist *.egg src/app/cli src/app/core src/app/common # Add any directories you want pytest to ignore
