+++ src/app/features/auth/security.py
@@ -1,0 +1,81 @@
+import logging
+from datetime import datetime, timedelta, timezone
+from typing import Optional, Annotated
+
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from jose import JWTError, jwt
+from pydantic import ValidationError
+import bcrypt
+
+from ...core.config import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES
+from . import schemas, service as auth_service
+from . import models
+
+logger = logging.getLogger(__name__)
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/auth/token")
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return bcrypt.checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))
+
+def get_password_hash(password: str) -> str:
+    hashed = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt())
+    return hashed.decode('utf-8')
+
+def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]) -> models.User:
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        sub: Optional[str] = payload.get("sub")
+        if sub is None:
+            logger.warning("Token sub (username) is missing.")
+            raise credentials_exception
+    except JWTError as e:
+        logger.error(f"JWT decoding error: {e}")
+        raise credentials_exception
+    except ValidationError as e:
+        logger.error(f"Token data validation error: {e}")
+        raise credentials_exception
+
+    user = await auth_service.get_user_by_username(username=sub)
+    if user is None:
+        logger.warning(f"User not found for username: {sub}")
+        raise credentials_exception
+    if not user.is_active:
+        logger.warning(f"User {sub} is inactive.")
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Inactive user")
+    return user
+
+async def get_current_active_user(current_user: Annotated[models.User, Depends(get_current_user)]) -> models.User:
+    return current_user
+
+async def get_current_active_admin_user(current_user: Annotated[models.User, Depends(get_current_user)]) -> models.User:
+    if current_user.role != "admin":
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="The user doesn't have enough privileges",
+        )
+    return current_user
+
+async def get_current_active_customer_user(current_user: Annotated[models.User, Depends(get_current_user)]) -> models.User:
+    if current_user.role != "customer":
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Operation not permitted for this user role.",
+        )
+    return current_user
