+++ src/app/features/inventory/schemas.py
