+++ src/app/features/inventory/tests/test_inventory.py
